pyproject.toml
[project]
name = "Mnemonic"
version = "1.0.0"
description = "Mémoire vectorielle locale pour agents autonomes"
authors = [{ name = "Ybucaille"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "numpy>=1.21",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0"
]

[project.optional-dependencies]
dev = ["pytest"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "."